refactor: repository 추상화 객체 추가
