+++ hub/src/main/java/com/klp/hub/inventory/domain/repository/InventoryRepository.java
@@ -1,0 +1,10 @@
+package com.klp.hub.inventory.domain.repository;
+
+import com.klp.hub.inventory.domain.Inventory;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface InventoryRepository {
+    Optional<Inventory> findByProductId(UUID productId);
+}
