+++ hub/src/main/java/com/klp/hub/inventory/infrastructure/repository/InventoryJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 import java.util.UUID;
 
-public interface InventoryRepository extends JpaRepository<Inventory, Long> {
+public interface InventoryJpaRepository extends JpaRepository<Inventory, Long> {
     Optional<Inventory> findByProductId(UUID productId);
 }
